+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ---
 
 - [Optimize:optimize spring value processor](https://github.com/Tencent/spring-cloud-tencent/pull/669)
+- [Optimize:optimize spring annotation.](https://github.com/Tencent/spring-cloud-tencent/pull/689)